astronomy: Duncan Steel
This adds Duncan Steel as an expert on Astronomy.

From [Wikipedia](https://en.wikipedia.org/wiki/Duncan_Steel):

*Duncan I. Steel FRAS (born 1955), is a British scientist born in
Midsomer Norton, Somerset. Currently he lives in Wellington, New
Zealand, but holds visiting positions as a Professor of Astrobiology at
the University of Buckingham in England; as a Space Scientist at
NASA-Ames Research Center in California; and as an Astronomer at Armagh
Observatory in Northern Ireland. Duncan is a space science authority who
has worked with NASA to assess the threat of comet and asteroid
collisions and investigate technologies to avert such impacts. He is
also the author of four popular-level science books on space, and
regularly writes for The Guardian and various other newspapers and
magazines. He is a discoverer of minor planets including the main-belt
asteroid 9767 Midsomer Norton.*